refactor: update home name to notes
