--- app/src/main/java/com/mutkuensert/highlightandnote/feature/note/presentation/home/HomeRoute.kt
@@ -1,6 +1,0 @@
-package com.mutkuensert.highlightandnote.feature.note.presentation.home
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class HomeRoute(val receivedText: String? = null)
