--- app/src/main/java/com/mutkuensert/highlightandnote/feature/note/presentation/home/NoteUiModel.kt
@@ -1,3 +1,0 @@
-package com.mutkuensert.highlightandnote.feature.note.presentation.home
-
-data class NoteUiModel(val id: Int, val text: String)
