+++ app/src/main/java/com/mutkuensert/highlightandnote/feature/note/presentation/notes/NoteUiModel.kt
@@ -1,0 +1,3 @@
+package com.mutkuensert.highlightandnote.feature.note.presentation.notes
+
+data class NoteUiModel(val id: Int, val text: String)
