+++ app/src/main/java/com/mutkuensert/highlightandnote/feature/note/presentation/notes/NotesRoute.kt
@@ -1,0 +1,6 @@
+package com.mutkuensert.highlightandnote.feature.note.presentation.notes
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class NotesRoute(val receivedText: String? = null)
